Implement DXYN (Draw) instruction using SDL2

+++ src/data.cpp
@@ -1,5 +1,5 @@
 #include "data.h"
 
 namespace info {
-	uint32_t SCALE_FACTOR = 20;
+	uint16_t SCALE_FACTOR = 20;
 }
